requirements.txt
fastapi
uvicorn
pillow
pytesseract
easyocr
transformers
python-multipart
